or1k_jtag_adv: try to not use polling on status
We now read more bits we are expected to have while
doing burst read without polling for status bit first.

Once we get our bits, we look for the status bit then
we just have to shift the rx buffer depending on the
status bit position.

I should be a bit faster and it fixes a problem with
the ftdi driver when doing DRSHIFT with an end state
equal to DRSHIFT.

TODO: Need some cleanup to completly remove the old
polling code.

Signed-off-by: Franck Jullien <franck.jullien@gmail.com>

Conflicts:

	src/target/or1k_jtag_adv.c