or1k_jtag_adv.c: don't pull waiting for status bit
We don't pull when waiting for the status bit to arrive.

The polling method was triggering an USB transfert at every
polling cycle.

We now shift some more bits, check if we can find the status
bit on the first few bits received in the buffer then we shift
the buffer to discard dummy bits before the status bit.

This method also solves a problem when using the ftdi driver.
Because of an unknown reason, a DRSHIFT command with an end
state equal to DRSHIFT (what we were doing when polling) generates
a TMS transition and a misbehavior in the adb_dbg_if. Now we do
a classic DRSHIFT ending in IDLE state.

Signed-off-by: Franck Jullien <franck.jullien@gmail.com>